+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#670](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/670))
 - `opentelemetry-instrumentation-redis` added request_hook and response_hook callbacks passed as arguments to the instrument method.
   ([#669](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/669))
-- `opentelemetry-instrumentation-botocore` add request_hooks and response_hooks
+- `opentelemetry-instrumentation-botocore` add `request_hook` and `response_hook` callbacks
   ([679](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/679))
 
 ### Changed